explicit cholesky for 3x3s
