+++ src/cones/coneops_powcone.jl
@@ -291,16 +291,11 @@ function higher_correction!(
     u = K.vec_work
     z = K.z
 
-    # lu factorization
-    getrf!(H,K.ws)
-    if K.ws.info[] == 0     # lu decomposition is successful
-        # @. u = ds
-        @inbounds for i = 1:3
-            u[i] = ds[i]
-        end
-        getrs!(H,K.ws,u)    # solve H*u = ds
-    else
-        # @. η = zero(T)
+    #solve H*u = ds
+    issuccess = cholesky_3x3_explicit_factor!(K.cholH,H)
+    if issuccess 
+        cholesky_3x3_explicit_solve!(u,K.cholH,ds)
+    else 
         @inbounds for i = 1:3
             η[i] = zero(T)
         end
